feat(WinningNumberValidator): 당첨 번호 검사 로직을 validator에서 수행하도록 수정
