+++ src/main/java/lotto/view/InputView.java
@@ -1,12 +1,9 @@
 package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-import lotto.info.LottoInfo;
 import lotto.message.ErrorMessage;
+import lotto.validator.WinningNumbersValidator;
 
 public class InputView {
     public static int getMoneyToBuy() {
@@ -41,42 +38,10 @@ public class InputView {
             try {
                 OutputView.notifyEnterWinningMoney();
                 String input = Console.readLine();
-                return validateWinningNumbers(input);
+                return WinningNumbersValidator.validateWinningNumbers(input);
             } catch (Exception e) {
                 OutputView.printErrorMessage(e.getMessage());
             }
-        }
-    }
-
-    public static List<Integer> validateWinningNumbers(String input) {
-        try {
-            String[] tokens = input.split(",");
-            if (tokens.length != LottoInfo.COUNT.getNumber()) {
-                throw new IllegalArgumentException(ErrorMessage.INVALID_COUNT_ERROR_MESSAGE.getMessage());
-            }
-
-            List<Integer> winningNumbers = new ArrayList<>();
-            Set<Integer> uniqueNumbers = new HashSet<>();
-            for (String s : tokens) {
-                String token = s.trim();
-
-                int number = Integer.parseInt(token);
-
-                if (number < 1 || number > 45) {
-                    throw new IllegalArgumentException(ErrorMessage.OUT_OF_RANGE_ERROR_MESSAGE.getMessage());
-                }
-
-                if (uniqueNumbers.contains(number)) {
-                    throw new IllegalArgumentException(ErrorMessage.DUPLICATE_NUMBER_ERROR_MESSAGE.getMessage());
-                }
-
-                uniqueNumbers.add(number);
-                winningNumbers.add(number);
-            }
-
-            return winningNumbers;
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException(ErrorMessage.NUMBER_FORMAT_ERROR_MESSAGE.getMessage(), e);
         }
     }
 }
